evaluators/presidio/pyproject.toml
[tool.poetry]
name = "langevals-presidio"
version = "0.1.0"
description = "LangEvals integration for Presidio"
authors = ["Rogerio Chaves <rogerio@langwatch.ai>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
langevals-core = { path = "../../langevals_core", develop = true }
presidio-analyzer = "^2.2.355"
presidio-anonymizer = "^2.2.355"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.2"
python-dotenv = "1.0.1"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
